Extract specific SDM goals and phases
